Separate cached discovery code (and testing)
A small reorganisation to separate

 - cachedDiscovery from namespacer, since it's used independently
 - things needed for testing the above, from non-testing code
